Added Login Persistence
After the user has logged in once to add new events, the program will
write a plist file that way the next time the app starts up it will
know that the user has logged in.  There are some problems with this
though….may need to change it in the future.